// src/store/general/types.ts
export enum GeneralActionTypes {
  INIT_APPLICATION = '[General] INIT_APPLICATION',
  INIT_APPLICATION_SUCCESS = '[General] INIT_APPLICATION_SUCCESS',
  INIT_APPLICATION_ERROR = '[General] INIT_APPLICATION_ERROR',
  CHANGE_LANGUAGE = '[General] CHANGE_LANGUAGE',
  CHANGE_LANGUAGE_SUCCESS = '[General] CHANGE_LANGUAGE_SUCCESS',
  CHANGE_LANGUAGE_FAIL = '[General] CHANGE_LANGUAGE_FAIL',
  GET_LOCALE = '[General] GET_LOCALE',
  GET_LOCALE_SUCCESS = '[General] GET_LOCALE_SUCCESS',
  GET_LOCALE_FAIL = '[General] GET_LOCALE_FAIL',
  GET_LOCATION = '[General] GET_LOCATION',
  GET_LOCATION_SUCCESS = '[General] GET_LOCATION_SUCCESS',
  GET_LOCATION_FAIL = '[General] GET_LOCATION_FAIL',
  GET_DEVICE_INFO = '[General] GET_DEVICE_INFO',
  SET_LOCATION = '[General] SET_LOCATION',
  UPDATE_SKELETON = '[General] UPDATE_SKELETON',
  UPDATE_LOADING = '[General] UPDATE_LOADING',
  SAVE_GTM_DATA = '[General] SAVE_GTM_DATA',

  GET_CALENDAR_DATA = '[General] GET_CALENDAR_DATA',
  GET_CALENDAR_DATA_SUCCESS = '[General] GET_CALENDAR_DATA_SUCCESS',
  GET_CALENDAR_DATA_FAIL = '[General] GET_CALENDAR_DATA_FAIL'
}